add metadata to vpipe module
