+++ scripts/vp_transformer.py
@@ -18,6 +18,7 @@ from sr2silo.lapis import submit
 from sr2silo.process import pair_normalize_reads
 from sr2silo.s3 import compress_bz2, upload_file_to_s3
 from sr2silo.translation import translate
+from sr2silo.vpipe.metadata import sample_id_decoder
 
 logging.basicConfig(
     level=logging.DEBUG, format="%(asctime)s - %(levelname)s - %(message)s"
@@ -35,31 +36,6 @@ def load_config(config_file: Path) -> dict:
     except json.JSONDecodeError as e:
         logging.error(f"Error decoding JSON from config file: {config_file} - {e}")
         raise
-
-
-def sample_id_decoder(sample_id: str) -> dict:
-    """Decode the sample ID into individual components.
-
-    Args:
-        sample_id (str): The sample ID to decode.
-
-    Returns:
-        dict: A dictionary containing the decoded components.
-              containing the following keys:
-                - sequencing_well_position (str : sequencing well position)
-                - location_code (int : code of the location)
-                - sampling_date (str : date of the sampling)
-    """
-    components = sample_id.split("_")
-    # Assign components to meaningful variable names
-    well_position = components[0]  # A1
-    location_code = components[1]  # 10
-    sampling_date = f"{components[2]}-{components[3]}-{components[4]}"  # 2024-09-30
-    return {
-        "sequencing_well_position": well_position,
-        "location_code": location_code,
-        "sampling_date": sampling_date,
-    }
 
 
 def batch_id_decoder(batch_id: str) -> dict:
