+++ src/sr2silo/__init__.py
@@ -5,4 +5,8 @@
 """sr2silo connects pairs, normalizes reads, and converts BAM to SAM files."""
 from __future__ import annotations
 
+import sr2silo.vpipe as vpipe
+
 __version__ = "0.0.2"
+
+__all__ = ["vpipe"]
