+++ src/sr2silo/vpipe/__init__.py
@@ -1,0 +1,4 @@
+"""Implements V-Pipe specific utilities.
+
+  i.e. extracting metadata from V-Pipe Filenaming Conventions.
+  """
